Create screens folder for restaurants feature
Add RestaurantsScreen to it, transfer code from App.js file to it.

Add more folder structure to the application
