+++ App.js
@@ -1,36 +1,13 @@
 import { StatusBar as ExpoStatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View, SafeAreaView, Platform, StatusBar } from 'react-native';
-import { Searchbar } from 'react-native-paper';
+
+import RestaurantsScreen from './src/features/restaurants/screens/RestaurantsScreen';
 
 const isAndroid = Platform.OS
 export default function App() {
   return (
     <>
-      <SafeAreaView style={styles.container}>
-        <View style={styles.searchContainer}>
-        <Searchbar placeholder='Search'/>
-        </View>
-        <View style={styles.listContainer}>
-          <Text>List</Text>
-        </View>
-      </SafeAreaView>
+      <RestaurantsScreen />
       <ExpoStatusBar />
     </>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: 'white',
-    paddingTop: Platform.OS === "android" ? StatusBar.currentHeight : 0
-  },
-  searchContainer: {
-    paddingVertical: 10,
-    paddingHorizontal: 20,
-  },
-  listContainer: {
-    backgroundColor: '#123',
-    flex: 1
-  }
-})
