+++ src/features/restaurants/screens/RestaurantsScreen.js
@@ -1,0 +1,34 @@
+import React from 'react';
+import { View, Text, SafeAreaView, StyleSheet, StatusBar } from 'react-native';
+import { Searchbar } from 'react-native-paper';
+
+const RestaurantsScreen = () => {
+  return (
+    <SafeAreaView style={styles.container}>
+      <View style={styles.searchContainer}>
+      <Searchbar placeholder='Search'/>
+      </View>
+      <View style={styles.listContainer}>
+        <Text>List</Text>
+      </View>
+    </SafeAreaView>
+  )
+}
+
+export default RestaurantsScreen;
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: 'white',
+    paddingTop: Platform.OS === "android" ? StatusBar.currentHeight : 0
+  },
+  searchContainer: {
+    paddingVertical: 10,
+    paddingHorizontal: 20,
+  },
+  listContainer: {
+    backgroundColor: '#123',
+    flex: 1
+  }
+})
